server/requirements.txt
# Core Django
Django>=5.1.2,<5.3
djangorestframework>=3.15.2,<4.0
djangorestframework-simplejwt>=5.3.1,<6.0

# Database
mysqlclient>=2.2.7,<3.0  # MySQL support (optional)

# CORS
django-cors-headers>=4.4.0,<5.0

# Environment variables
python-decouple>=3.8,<4.0

# Google OAuth2
google-auth>=2.35.0,<3.0
requests>=2.31.0,<3.0

# Security & Authentication
PyJWT>=2.9.0,<3.0
argon2-cffi>=23.1.0,<25.0
pyotp>=2.9.0,<3.0
qrcode>=7.4.2,<9.0
Pillow>=10.4.0,<13.0
cryptography>=42.0.0,<48.0

# Static files
whitenoise>=6.7.0,<7.0

# Production server (ASGI)
uvicorn[standard]>=0.32.1,<1.0

# Optional: Production database adapter
# psycopg2-binary>=2.9.9,<3.0  # PostgreSQL support (uncomment if needed)

# Optional: Caching
# redis>=5.0.0,<6.0  # Redis support (uncomment if needed)
# django-redis>=5.4.0,<6.0  # Django Redis cache backend (uncomment if needed)
